Move watcher from pkg/client/userd/k8s to pkg/k8sapi
So that it can be reused in the traffic-manager
