+++ pkg/client/userd/trafficmgr/workloads.go
@@ -16,7 +16,6 @@ import (
 	"k8s.io/client-go/tools/cache"
 
 	"github.com/datawire/dlib/dlog"
-	"github.com/telepresenceio/telepresence/v2/pkg/client/userd/k8s"
 	"github.com/telepresenceio/telepresence/v2/pkg/k8sapi"
 )
 
@@ -33,8 +32,8 @@ const statefulsets = 2
 
 // namespacedWASWatcher is watches Workloads And Services (WAS) for a namespace
 type namespacedWASWatcher struct {
-	svcWatcher *k8s.Watcher
-	wlWatchers [3]*k8s.Watcher
+	svcWatcher *k8sapi.Watcher
+	wlWatchers [3]*k8sapi.Watcher
 }
 
 // svcEquals compare only the Service fields that are of interest to Telepresence. They are
@@ -144,11 +143,11 @@ func newNamespaceWatcher(c context.Context, namespace string, cond *sync.Cond) *
 	ki := k8sapi.GetK8sInterface(c)
 	appsGetter := ki.AppsV1().RESTClient()
 	w := &namespacedWASWatcher{
-		svcWatcher: k8s.NewWatcher("services", namespace, ki.CoreV1().RESTClient(), &core.Service{}, cond, svcEquals),
-		wlWatchers: [3]*k8s.Watcher{
-			k8s.NewWatcher("deployments", namespace, appsGetter, &apps.Deployment{}, cond, workloadEquals),
-			k8s.NewWatcher("replicasets", namespace, appsGetter, &apps.ReplicaSet{}, cond, workloadEquals),
-			k8s.NewWatcher("statefulsets", namespace, appsGetter, &apps.StatefulSet{}, cond, workloadEquals),
+		svcWatcher: k8sapi.NewWatcher("services", namespace, ki.CoreV1().RESTClient(), &core.Service{}, cond, svcEquals),
+		wlWatchers: [3]*k8sapi.Watcher{
+			k8sapi.NewWatcher("deployments", namespace, appsGetter, &apps.Deployment{}, cond, workloadEquals),
+			k8sapi.NewWatcher("replicasets", namespace, appsGetter, &apps.ReplicaSet{}, cond, workloadEquals),
+			k8sapi.NewWatcher("statefulsets", namespace, appsGetter, &apps.StatefulSet{}, cond, workloadEquals),
 		},
 	}
 	return w
@@ -274,7 +273,7 @@ func (w *workloadsAndServicesWatcher) addNSLocked(c context.Context, ns string) 
 	nw := newNamespaceWatcher(c, ns, &w.cond)
 	w.nsWatchers[ns] = nw
 	for _, l := range w.nsListeners {
-		nw.svcWatcher.AddStateListener(&k8s.StateListener{Cb: l})
+		nw.svcWatcher.AddStateListener(&k8sapi.StateListener{Cb: l})
 	}
 	return nw
 }
@@ -309,7 +308,7 @@ func (w *workloadsAndServicesWatcher) addActiveNamespaceListener(l func()) {
 	w.Lock()
 	w.nsListeners = append(w.nsListeners, l)
 	for _, nw := range w.nsWatchers {
-		nw.svcWatcher.AddStateListener(&k8s.StateListener{Cb: l})
+		nw.svcWatcher.AddStateListener(&k8sapi.StateListener{Cb: l})
 	}
 	w.Unlock()
 }
